Rename FracAddCheckProver to FracAddCircuit and move it to circuit.rs
The type holds Vec<Fraction<FieldVec>> and an allocator, and its
substantial method materializes every layer of the GKR circuit from the
witness. It proves nothing. The things that prove are LayerProver and the
driver functions. A type whose job is to hold the materialized layers, with
a name ending in Prover, sends a reader looking for round logic that is not
there.

Its `new` becomes `build`, because it does not return Self: it returns the
circuit beside its root layer. Rust reserves `new` for constructors
returning Self, and this is the case that convention exists to catch.

Both move to fracaddcheck/circuit.rs together with n_layers, pop_layer,
prove and prove_layers, since those are the circuit and driving it end to
end. reduce_layer, BatchProveOutput, combine_claims and
batch_prove_unequal_depths stay in mod.rs, untouched apart from the type
name in one signature and one map.

The module is private with a re-export, so fracaddcheck::FracAddCircuit is
the one canonical path and there is no deeper path to drift to. That
matches how LayerProver and unpad_leaf_claim are already surfaced.

The transcript is byte-identical. A throwaway harness dumped
transcript.finalize() plus the returned eval_point and fractions for depths
[2,4,5], [4,4,4] and [0,3] before and after; both files hash to
83cc802a27a70475b741e6d783d38013c98b57caab29377a395f022feaf814b3.

The bench's group id moves from "fracaddcheck/new" to "fracaddcheck/build"
so it agrees with the method it measures. That orphans any stored criterion
baseline for it, which is the right trade while no performance work depends
on one.

Five doc paragraphs that wrapped a sentence across /// lines are rewritten.
Four are prose carried over unchanged for months; one is from the pop_layer
doc added recently. A new file should not ship with them.
